IBH-89 Add the HTTPClientHelper(null) to resolve the issue
The code to address this issue existed in an earlier version when the
configuration Since we have a singleton HubBambooUtils use that object
to only initialize the Restlet engine with the correct client to resolve
the problem.  The use of the singleton is needed to only add the client
once when the plugin is run.